requirements.txt
aiogram>=3.4.1
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.29
asyncpg>=0.29.0
cryptography>=42.0.5
python-dotenv>=1.0.1
qdrant-client>=1.8.0
fastembed>=0.2.6
pdfplumber>=0.11.0
python-docx>=1.1.0
langchain-text-splitters>=0.0.1
openai>=1.14.0